+++ src/main/java/pro/sky/petshelterbot/controllers/ShelterReportController.java
@@ -6,7 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pro.sky.petshelterbot.entity.Pet;
 import pro.sky.petshelterbot.entity.Report;
-import pro.sky.petshelterbot.service.ShelterReportService;
+import pro.sky.petshelterbot.service.ReportService;
 
 import java.util.List;
 
@@ -15,10 +15,10 @@ import java.util.List;
 @Tag(name = "ShelterReportController")
 public class ShelterReportController {
 
-    private final ShelterReportService shelterReportService;
+    private final ReportService reportService;
 
-    public ShelterReportController(ShelterReportService shelterReportService) {
-        this.shelterReportService = shelterReportService;
+    public ShelterReportController(ReportService reportService) {
+        this.reportService = reportService;
     }
 
     @GetMapping(path = "/{id}")
@@ -27,13 +27,13 @@ public class ShelterReportController {
             @PathVariable Long id,
             @RequestParam(defaultValue = "0") Integer pageNo,
             @RequestParam(defaultValue = "10") Integer pageSize) {
-        return ResponseEntity.ok(shelterReportService.findAllByPetId(id, pageNo, pageSize));
+        return ResponseEntity.ok(reportService.findAllByPetId(id, pageNo, pageSize));
     }
 
     @GetMapping("/overdue")
     @ApiResponse(description = "Возвращает животных на пробном периоде адоптации, для которых адоптеры не прислали текущие отчёты своевременно.")
     public ResponseEntity<List<Pet>> findOverdueReports() {
-        return ResponseEntity.ok(shelterReportService.findOverdueReports());
+        return ResponseEntity.ok(reportService.findOverdueReports());
     }
 
     @GetMapping("/all/{shelterId}")
@@ -45,7 +45,7 @@ public class ShelterReportController {
             @RequestParam(defaultValue = "0") Integer pageNo,
             @RequestParam(defaultValue = "10") Integer pageSize
     ) {
-        return ResponseEntity.ok(shelterReportService.findAllReportsByShelterId(shelterId, pageNo, pageSize));
+        return ResponseEntity.ok(reportService.findAllReportsByShelterId(shelterId, pageNo, pageSize));
     }
 
     @PutMapping()
@@ -55,7 +55,7 @@ public class ShelterReportController {
             "то пользователю отсылается сообщение с рекомендацией" +
             "заполнять отчёты более полно.")
     public ResponseEntity<Report> updateReport(@RequestBody Report report) {
-        return ResponseEntity.ok(shelterReportService.updateReport(report));
+        return ResponseEntity.ok(reportService.updateReport(report));
     }
 
     @GetMapping("/to-review")
@@ -64,6 +64,6 @@ public class ShelterReportController {
             @RequestParam(defaultValue = "0") Integer pageNo,
             @RequestParam(defaultValue = "10") Integer pageSize
     ) {
-        return ResponseEntity.ok(shelterReportService.findAllReportsToReview(pageNo, pageSize));
+        return ResponseEntity.ok(reportService.findAllReportsToReview(pageNo, pageSize));
     }
 }
