+++ src/main/java/pro/sky/petshelterbot/service/ReportService.java
@@ -6,7 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import pro.sky.petshelterbot.entity.Pet;
 import pro.sky.petshelterbot.entity.Report;
-import pro.sky.petshelterbot.exceptions.ShelterException;
 import pro.sky.petshelterbot.repository.PetRepository;
 import pro.sky.petshelterbot.repository.ReportRepository;
 
@@ -14,12 +13,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Service
-public class ShelterReportService {
+public class ReportService {
 
     private final ReportRepository reportRepository;
     private final PetRepository petRepository;
 
-    public ShelterReportService(ReportRepository reportRepository, PetRepository petRepository) {
+    public ReportService(ReportRepository reportRepository, PetRepository petRepository) {
         this.reportRepository = reportRepository;
         this.petRepository = petRepository;
     }
